Moved the annotation to a different facet to avoid circular dependencies and generally improve code structure
