+++ application/build.gradle
@@ -43,6 +43,7 @@ dependencies {
     implementation 'org.jetbrains:annotations:23.0.0'
 
     implementation project(':database')
+    implementation project(':utils')
 
     implementation 'net.dv8tion:JDA:5.0.0-alpha.9'
