+++ application/src/main/java/org/togetherjava/tjbot/commands/componentids/package-info.java
@@ -14,6 +14,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.commands.componentids;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
