+++ application/src/main/java/org/togetherjava/tjbot/commands/mathcommands/wolframalpha/api/package-info.java
@@ -7,6 +7,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.commands.mathcommands.wolframalpha.api;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
