+++ application/src/main/java/org/togetherjava/tjbot/commands/moderation/package-info.java
@@ -6,6 +6,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.commands.moderation;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
