+++ application/src/main/java/org/togetherjava/tjbot/commands/moderation/temp/package-info.java
@@ -5,6 +5,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.commands.moderation.temp;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
