+++ application/src/main/java/org/togetherjava/tjbot/config/package-info.java
@@ -6,6 +6,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.config;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
