+++ application/src/main/java/org/togetherjava/tjbot/package-info.java
@@ -5,6 +5,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
