+++ application/src/main/java/org/togetherjava/tjbot/routines/package-info.java
@@ -9,6 +9,6 @@
 @ParametersAreNonnullByDefault
 package org.togetherjava.tjbot.routines;
 
-import org.togetherjava.tjbot.misc.MethodsReturnNonnullByDefault;
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;
