+++ database/build.gradle
@@ -9,5 +9,7 @@ dependencies {
     implementation "org.xerial:sqlite-jdbc:${sqliteVersion}"
     implementation 'org.flywaydb:flyway-core:9.2.0'
     implementation 'org.jooq:jooq:3.17.2'
+
+    implementation project(':utils')
 }
