+++ utils/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'java'
+}
+dependencies {
+    implementation 'com.google.code.findbugs:jsr305:3.0.2'
+}
+
