+++ utils/src/main/java/org/togetherjava/tjbot/annotations/MethodsReturnNonnullByDefault.java
@@ -1,4 +1,4 @@
-package org.togetherjava.tjbot.misc;
+package org.togetherjava.tjbot.annotations;
 
 import javax.annotation.Nonnull;
 import javax.annotation.meta.TypeQualifierDefault;
